+++ server/src/app/controllers/flowController.js
@@ -87,43 +87,6 @@ const deleteFlow = async(req, res) => {
     });
 }
 
-const createMapFlow = async (req, res) => {
-  try {
-    const {data, userToken, flowName} = req.body
-
-    const query = { userId: userToken };
-
-    const update = {
-      $set: {
-        'flowList.$[flow].flow': {
-          nodes: data.nodes,
-          edges: data.edges,
-          viewport: data.viewport
-        }
-      }
-    };
-
-    const options = {
-      arrayFilters: [{ 'flow.name': flowName }] 
-    };
-
-    await User.findOneAndUpdate(query, update, options);
-
-    return res.status(200)
-  } catch (error) {
-    console.log(error)
-    res.status(500).json({message: error.message})
-  }
-}
-
-const getFlowMap = async (req, res) => {
-  try {
-    
-  } catch (error) {
-    console.log(error.message);
-    res.status(500).json({message: error.message})
-  }
-}
 
 module.exports = {
     newFlow,
@@ -131,5 +94,4 @@ module.exports = {
     getOneFlow,
     deleteFlow,
     updateFlow,
-    createMapFlow
 }
